# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bilemo",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://bilemo.com/api",
            "description": "Bilemo API"
        }
    ],
    "paths": {
        "/customers": {
            "get": {
                "tags": [
                    "customer"
                ],
                "description": "Get all customers.",
                "operationId": "App\\Controller\\CustomerController::collection",
                "responses": {
                    "200": {
                        "description": "Customers liste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenNotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "post": {
                "tags": [
                    "customer"
                ],
                "description": "Create new customer.",
                "operationId": "App\\Controller\\CustomerController::post",
                "requestBody": {
                    "request": "CreatePost",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "firstName",
                                    "lastName"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/customers/{id}": {
            "get": {
                "tags": [
                    "customer"
                ],
                "description": "Get customer by ID",
                "operationId": "App\\Controller\\CustomerController::item",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer detaille",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidID"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenNotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "customer"
                ],
                "description": "Delete customer by ID",
                "operationId": "App\\Controller\\CustomerController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidID"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenNotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "description": "Get all products.",
                "operationId": "App\\Controller\\ProductController::collection",
                "responses": {
                    "200": {
                        "description": "Products liste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenNotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "product"
                ],
                "description": "Get product by ID",
                "operationId": "App\\Controller\\ProductController::item",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detaille",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidID"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenNotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "registeredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "reference": {
                        "type": "string"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "camera": {
                        "type": "string"
                    },
                    "screenSize": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "NotFound": {
                "description": "This resource does not exist.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "This resource does not exist."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "InvalidID": {
                "description": "Invalid ID supplied",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Invalid ID supplied"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "TokenNotFound": {
                "description": "Token not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "JWT Token not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Ressource ID",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    }
}